Consolidate references to redis client
