+++ app/dashboard/session.js
@@ -2,7 +2,8 @@ const config = require("config");
 const guid = require("helper/guid");
 const session = require("express-session");
 const Store = require("connect-redis")(session);
-const redis = require("redis").createClient();
+const redis = require("models/redis");
+const client = new redis();
 
 // Session settings. It is important that session
 // comes before the cache so we know what to serve
@@ -17,12 +18,12 @@ module.exports = session({
   cookie: {
     httpOnly: true, // prevent the cookie's exposure to client-side js
     secure: true, // ensure the cookie is only accesible over HTTPS
-    domain: '', // prevent the cookie's exposure to sub domains
+    domain: "", // prevent the cookie's exposure to sub domains
     sameSite: true, // prevent the cookie's exposure to other sites
     maxAge: 1000 * 60 * 60 * 24 * 30, // 30 days in ms
   },
   store: new Store({
-    client: redis,
+    client,
     port: config.redis.port,
   }),
 });
