+++ app/scheduler/backup.js
@@ -1,7 +1,8 @@
 var fs = require("fs-extra");
 var config = require("config");
 var Moment = require("moment");
-var redis = require("redis").createClient();
+var redis = require("models/redis");
+var client = new redis();
 var tmp = require("helper/tempDir")();
 var ensure = require("helper/ensure");
 var encrypt = require("helper/encrypter").encrypt;
